Prefetch on file-load and save anime on main.lua
